Move state property to view from model because it can't work outside the view accotding to docs. And it is still doesn't work if it is set outside of view body.
