+++ catowser/catowser/BrowserContent/BrowserContentView.swift
@@ -7,12 +7,31 @@
 //
 
 import SwiftUI
+import CoreBrowser
+import JSPlugins
+
+/**
+ To avoid this, always declare state as private,
+ and place it in the highest view in the view hierarchy that needs access to the value.
+ https://developer.apple.com/documentation/swiftui/state
+ 
+ Sooo, can't use state properties in model types, WTF???
+ */
 
 struct BrowserContentView: View {
     @EnvironmentObject var model: BrowserContentModel
+    /// view state
+    @State private var contentType: Tab.ContentType = DefaultTabProvider.shared.contentState
+    /// Not initialized, will be initialized after `TabsListManager`
+    /// during tab opening. Used only during tab opening for optimization
+    private var previousTabContent: Tab.ContentType?
+    
+    init() {
+        TabsListManager.shared.attach(self)
+    }
     
     var body: some View {
-        switch model.contentType {
+        switch contentType {
         case .blank:
             EmptyView()
                 .background(.white)
@@ -25,6 +44,43 @@ struct BrowserContentView: View {
         default:
             EmptyView()
         }
-        
     }
 }
+
+extension BrowserContentView: TabsObserver {
+    func tabDidSelect(index: Int, content: Tab.ContentType, identifier: UUID) {
+        if let previousValue = previousTabContent, previousValue.isStatic && previousValue == content {
+            // Optimization to not do remove & insert of the same static view
+            return
+        }
+        contentType = content
+    }
+    
+    func tabDidReplace(_ tab: Tab, at index: Int) {
+        contentType = tab.contentType
+    }
+}
+
+#if DEBUG
+
+// https://martinlasek.medium.com/swiftui-understanding-binding-8e20269a76bc
+
+class DummyJSPluginsSource: JSPluginsSource {
+    typealias Program = JSPluginsProgramImpl
+    var pluginsProgram: Program {
+        JSPluginsProgramImpl([])
+    }
+}
+
+struct BrowserContentView_Previews: PreviewProvider {
+    static let model: BrowserContentModel = {
+        let source: DummyJSPluginsSource = .init()
+        return .init(source)
+    }()
+    
+    static var previews: some View {
+        EmptyView()
+    }
+}
+
+#endif
